requirements.txt
discord==1.7.3
dislash.py==1.4.9
discord.py==1.7.3
